# api/api-docs.yaml
openapi: 3.0.1
info:
  title: User API
  description: Documentation of the User API v1.0
  version: "1.0"
servers:
- url: http://localhost:9080/
  description: Generated server url
paths:
  /users/{userId}:
    delete:
      tags:
      - user-controller
      summary: Delete user
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - user-controller
      summary: Update user
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdateResponse'
  /users:
    get:
      tags:
      - user-controller
      summary: List users
      operationId: listUsers
      parameters:
      - name: pageNo
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagingResponseUser'
    post:
      tags:
      - user-controller
      summary: Create user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
components:
  schemas:
    UserUpdateRequest:
      required:
      - email
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    UserUpdateResponse:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    PagingResponseUser:
      required:
      - items
      - total
      type: object
      properties:
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
    User:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    UserCreateRequest:
      required:
      - email
      - firstName
      - lastName
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
    UserCreateResponse:
      required:
      - email
      - firstName
      - id
      - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string